--- java-playground/src/main/java/org/gonza/javaplayground/lotto/LottoNumbers.java
@@ -1,4 +1,0 @@
-package org.gonza.javaplayground.lotto;
-
-public class LottoNumbers {
-}
