+++ java-playground/src/main/java/org/gonza/javaplayground/lotto/domain/LottoNumbers.java
@@ -1,0 +1,4 @@
+package org.gonza.javaplayground.lotto.domain;
+
+public class LottoNumbers {
+}
